Code Review Suggestion 2
Since the navbar here is pretty much the same as the navbar in the navbar.js and Game_Signin.js, it would make it easier to debug and update the components in all pages at the same time if you just use the component implemented as a separated module. Even if they are slightly different, you can still control the rendering process with props.